requirements.txt
pdfminer3k
ply